requirements.txt
Scrapy==2.8.0
playwright==1.31.0
scrapy-playwright==0.0.26
scrapyd==1.4.1
scrapyd-client==1.2.3
beautifulsoup4==4.11.2
pygsheets==2.0.6
pandas==1.5.3
google-cloud-bigquery==3.6.0

